requirements.txt
Django==3.2.7
psycopg2-binary>=2.8
djangorestframework==3.12.2
django-cors-headers==3.5.0
pika==1.1.0
asgiref==3.3.1
sqlparse==0.4.1
